requirements.txt
Django~=4.2.7
djangorestframework~=3.14.0
pandas~=2.2.2
django-filter~=23.4
corsheaders